Fix: Add line break when commit is only one line
Update the logic to append the Jira info so a new line is added in front
of it when there is more than 0 lines. Previously this was more than 1
line, however, this causes an issue with merges and other single line
commits as there is no blank line after the summary.