Fix: Only git start lines if found in message
When appending the Jira info to the commit message, only add back in the
GIT_START_LINES if the lines was found during the enumeration. This
means it will only be printed if it was originally in the commit message.
This will fix a bug where certain git tools such as SourceTree apply the
pre-commit hook after the user has written the message, and as a result
printing the start line out would append it to the final commit, despite
the line starting with '#'.