Fix: Add blank line when commit is a single line
Update the logic to append the Jira info so an extra blank line is added
in front of it if the commit is a single line. Previously this only
added one line line, however, this causes an issue with merges and other
single line commits as there is no blank line after the summary.